+++ src/sevenmark/parser/bracket/bracket_media.rs
@@ -39,5 +39,6 @@ pub fn bracket_media_parser(parser_input: &mut ParserInput) -> Result<SevenMarkE
         location: Location { start, end },
         parameters: parameters.unwrap_or_default(),
         content: parsed_content.unwrap_or_default(),
+        resolved_info: None,
     }))
 }